add support for password verification
* restructured command line into a operation flow
* implemented `authenticated?` on PasswordFile
* implemented --verify command line options using `authenticated?`

+++ lib/htauth/cli/passwd.rb
@@ -33,8 +33,8 @@ module HTAuth
           @options.show_version   = false
           @options.show_help      = false
           @options.username       = nil
-          @options.delete_entry   = false
           @options.password       = ""
+          @options.operation      = []
         end
         @options
       end
@@ -77,6 +77,7 @@ Usage:
 
             op.on("-c", "--create", "Create a new file; this overwrites an existing file.") do |c|
               options.file_mode = HTAuth::File::CREATE
+              options.operation << :add_or_update
             end
 
             op.on("-d", "--crypt", "Force CRYPT encryption of the password.") do |c|
@@ -84,7 +85,7 @@ Usage:
             end
 
             op.on("-D", "--delete", "Delete the specified user.") do |d|
-              options.delete_entry = d
+              options.operation << :delete
             end
 
             op.on("-h", "--help", "Display this help.") do |h|
@@ -102,6 +103,7 @@ Usage:
             op.on("-n", "--stdout", "Do not update the file; Display the results on stdout instead.") do |n|
               options.send_to_stdout = true
               options.passwdfile     = HTAuth::File::STDOUT_FLAG
+              options.operation     << :stdout
             end
 
             op.on("-p", "--plaintext", "Do not encrypt the password (plaintext).") do |p|
@@ -114,6 +116,10 @@ Usage:
 
             op.on("-v", "--version", "Show version info.") do |v|
               options.show_version = v
+            end
+
+            op.on("--verify", "Verify password for the specified user") do |v|
+              options.operation << :verify
             end
 
             op.separator ""
@@ -138,8 +144,9 @@ Usage:
         begin
           option_parser.parse!(argv)
           show_version if options.show_version
-          show_help if options.show_help 
+          show_help if options.show_help
 
+          raise ::OptionParser::ParseError, "only one of --create, --stdout, --verify, --delete may be specified" if options.operation.size > 1
           raise ::OptionParser::ParseError, "Unable to send to stdout AND create a new file" if options.send_to_stdout and (options.file_mode == File::CREATE)
           raise ::OptionParser::ParseError, "a username is needed" if options.send_to_stdout and argv.size < 1
           raise ::OptionParser::ParseError, "a username and password are needed" if options.send_to_stdout and options.batch_mode  and ( argv.size < 2 ) 
@@ -147,6 +154,7 @@ Usage:
           raise ::OptionParser::ParseError, "a passwordfile and username are needed" if argv.size < 2
           raise ::OptionParser::ParseError, "options -i and -b are mutually exclusive" if options.batch_mode && options.read_stdin_once
 
+          options.operation  = options.operation.shift || :add_or_update
           options.passwdfile = argv.shift unless options.send_to_stdout
           options.username   = argv.shift
           options.password   = argv.shift if options.batch_mode
@@ -158,36 +166,60 @@ Usage:
         end
       end
 
+      def fetch_password(width=20)
+        return options.password if options.batch_mode
+        console = Console.new
+        if options.read_stdin_once then
+          pw_in = console.read_answer
+          return pw_in
+        end
+
+        case options.operation
+        when :verify
+          pw_in = console.ask("Enter password: ".rjust(width))
+          raise PasswordError, "password '#{pw_in}' too long" if pw_in.length >= MAX_PASSWD_LENGTH
+        when :add_or_update
+          pw_in = console.ask("New password: ".rjust(width))
+          raise PasswordError, "password '#{pw_in}' too long" if pw_in.length >= MAX_PASSWD_LENGTH
+
+          pw_validate = console.ask("Re-type new password: ".rjust(width))
+          raise PasswordError, "They don't match, sorry." unless pw_in == pw_validate
+        end
+
+        return pw_in
+      end
+
       def run(argv, env = ENV)
         begin
           parse_options(argv)
+          console = Console.new
           passwd_file = PasswdFile.new(options.passwdfile, options.file_mode)
-
-          if options.delete_entry then
+          case options.operation
+          when :delete
             passwd_file.delete(options.username)
-          else
-            action = passwd_file.has_entry?(options.username) ? "Changing" : "Adding"
-            if !options.batch_mode then
-              console = Console.new
-
-              if options.read_stdin_once then
-                pw_in = console.read_answer
+            passwd_file.save!
+          when :verify
+            if passwd_file.has_entry?(options.username) then
+              pw_in = fetch_password
+              if passwd_file.authenticated?(options.username, pw_in) then
+                $stderr.puts "Password for user #{options.username} correct."
               else
-                console.say "#{action} password for #{options.username}."
-
-                pw_in       = console.ask("        New password: ")
-                raise PasswordError, "password '#{pw_in}' too long" if pw_in.length >= MAX_PASSWD_LENGTH
-
-                pw_validate = console.ask("Re-type new password: ")
-                raise PasswordError, "They don't match, sorry." unless pw_in == pw_validate
+                raise HTAuth::Error, "Password verification for user #{options.username} failed."
               end
-              options.password = pw_in
+            else
+              raise HTAuth::Error, "User #{options.username} not found"
             end
+          when :add_or_update
+            options.password = fetch_password
+            action = passwd_file.has_entry?(options.username) ? "Changing" : "Adding"
+            console.say "#{action} password for #{options.username}."
             passwd_file.add_or_update(options.username, options.password, options.algorithm, options.algorithm_args)
+            passwd_file.save!
+          when :stdout
+            options.password = fetch_password
+            passwd_file.add_or_update(options.username, options.password, options.algorithm, options.algorithm_args)
+            passwd_file.save!
           end
-
-          passwd_file.save! 
-
         rescue HTAuth::FileAccessError => fae
           msg = "Password file failure (#{options.passwdfile}) "
           $stderr.puts "#{msg}: #{fae.message}"
